🔄 Break job queue circular deps by splitting transport failure layers
Extract shared constants and separate DB terminalization from enqueue follow-ups so mock queue providers no longer import the full run/enqueue cycle.
