+++ packages/scms-server/src/backend/jobs/jobQueueConstants.server.ts
@@ -1,0 +1,2 @@
+/** Max consumer delivery attempts before terminalizing via handleTransportFailure. */
+export const MAX_JOB_QUEUE_DELIVERY_ATTEMPTS = 3;
